Move whitelist to a separate file. Modify grunt task to concat the files together. Fix a bug with the public API that didn't pass through an argument properly.
