+++ src/js/videojs-dash.js
@@ -2,268 +2,7 @@
 var window_ = require('global/window');
 var videojs = require('video.js');
 var dashjs = require('dashjs');
-
-  var WhitelistPlugin = function () {
-    var globalManifest;
-    var globalDashManifestModel;
-    var globalStreamInfo;
-    var globalPlayer;
-    var globalAdapter;
-    var globalRulesController;
-    var globalOgPlayerSetQuality;
-
-    var whiteList;
-    var currentPeriodIndex;
-    var selectorFunction;
-
-    /**
-     * Custom object overriding the RulesController, specifically the applyRules
-     * function to accomodate the whitelisting of Representations
-     */
-    var CustomRulesController = function() {
-      globalManifest = this.factory.getSingletonInstance(this.context, 'ManifestModel');
-      globalDashManifestModel = this.factory.getSingletonInstance(this.context,
-                                                                  'DashManifestModel');
-      globalAdapter = this.factory.getSingletonInstance(this.context, 'DashAdapter');
-      globalRulesController = this.parent;
-      var ogApplyRules = this.parent.applyRules;
-      return {
-        applyRules: function(rulesArr, streamProcessor, callback, current, overrideFunc) {
-          var newCallbackFunc = function(result) {
-            if (typeof result.value !== 'object') {
-              result.value = whitelistImplementer(current, result.value, streamProcessor);
-            }
-
-              callback(result);
-            };
-            ogApplyRules(rulesArr, streamProcessor, newCallbackFunc, current, overrideFunc);
-        }
-      };
-    };
-
-    function periodSwitch(e) {
-      globalStreamInfo = e.toStreamInfo;
-      currentPeriodIndex = e.toStreamInfo.index;
-      whiteList = undefined;
-    }
-
-    /**
-     * @param  {number} - current index of quality
-     * @param  {number} - proposed index of quality
-     * @param  {StreamProcessor} - StreamProcessor object
-     * @return {number} - index of quality selected
-     */
-    function whitelistImplementer(current, proposed, streamProcessor) {
-      if(current === 0 || proposed === 0) {
-        return proposed;
-      }
-      if (current === proposed) {
-        return proposed;
-      }
-      if (whiteList) {
-        var manifest = Object.assign({}, globalManifest.getValue());
-        var mediaInfo = streamProcessor.getMediaInfo();
-        var adaptationSet = globalAdapter.getDataForMedia(mediaInfo);
-        adaptationSet = globalDashManifestModel.getAdaptationForIndex(
-                          adaptationSet.index,
-                          manifest,
-                          currentPeriodIndex);
-        var representations = adaptationSet.Representation_asArray;
-        var proposedRepresentation = representations[proposed-1];
-        var whiteListFilterFunc = whiteList[adaptationSet.id];
-        if (whiteListFilterFunc) {
-          var filteredReps = representations.filter(whiteListFilterFunc);
-          for (var i=0; i < filteredReps.length; i++) {
-            if (filteredReps[i].id === proposedRepresentation.id) {
-              return proposed;
-            }
-          }
-          return selectorFunction(proposed, adaptationSet, filteredReps);
-        }
-      }
-      return proposed;
-    }
-
-    /**
-     * @param  {number} - proposed index of quality (nonzero based)
-     * @param  {Representation} - Representation object of proposed quality
-     * @param  {Array} - Array of Representations representing the whitelist available
-     * @return {number} - index of quality to be applied
-     */
-    function defaultSelector(proposedQuality, adaptation, whiteListSet) {
-      if (whiteListSet && whiteListSet.length > 0) {
-        var proposedRepresentation = adaptation.Representation_asArray[proposedQuality];
-        var i,
-            len,
-            bestMatchedRep;
-        whiteListSet.sort(function(a, b) {
-          return a.bandwidth - b.bandwidth;
-        });
-        for (i=0, len=whiteListSet.length; i<len; i++) {
-          var before = whiteListSet[i-1] ? whiteListSet[i-1].bandwidth : 0;
-          var current = whiteListSet[i].bandwidth;
-          var max = whiteListSet[len - 1].bandwidth;
-          if (max <= proposedRepresentation.bandwidth) {
-            bestMatchedRep = whiteListSet[len-1];
-            break;
-          }
-          if (proposedRepresentation.bandwidth > before &&
-              proposedRepresentation.bandwidth <= current) {
-            var beforeDelta = proposedRepresentation.bandwidth - before;
-            var afterDelta = current - proposedRepresentation.bandwidth;
-            var index;
-            if (beforeDelta < afterDelta) {
-              index = i = 0 ? 1 : i + 1;
-              bestMatchedRep = whiteListSet[index];
-              break;
-            }
-            else {
-              bestMatchedRep = whiteListSet[i];
-              break;
-            }
-          }
-        }
-        var a,
-          length;
-        for (a=0, length=adaptation.Representation_asArray.length; a < length; a++ ){
-          if (adaptation.Representation_asArray[a].id === bestMatchedRep.id) {
-            return a;
-          }
-        }
-
-        return 0;
-      }
-      return proposedQuality;
-    }
-
-    selectorFunction = defaultSelector;
-
-    return {
-      /**
-       * Sets the selector function to use when choosing a bitrate.
-       * When Dash.js attempts to go to a bitrate that isn't in the whitelist,
-       * the selector function is used to determine which bitrate to go to
-       * instead.
-       *
-       * @param {[type]} sFunc - The function to use when selecting whitelisted bitrates.
-       */
-      setSelector: function (sFunc) {
-        selectorFunction = sFunc;
-      },
-      /**
-       * Initialize plugin.
-       *
-       * @param  {MediaPlayer} player - Dash MediaPlayer
-       */
-      initialize: function (player) {
-        this.initialized = true;
-        player.extend('RulesController', CustomRulesController, true);
-        player.on(dashjs.MediaPlayer.events.PERIOD_SWITCH_COMPLETED, periodSwitch);
-        globalPlayer = player;
-
-        globalOgPlayerSetQuality = player.setQualityFor;
-
-        // Patch the player object so that the original setQualityFor cannot be used.
-        player.setQualityFor = this.setQualityFor;
-      },
-      /**
-       * Returns all adaptations for the currently playing period.
-       * @return {Array} - Array of Representations
-       */
-      getAdaptations: function() {
-        if (globalManifest) {
-          var manifest = Object.assign({}, globalManifest.getValue());
-          return manifest.Period_asArray[currentPeriodIndex].AdaptationSet_asArray;
-        }
-      },
-      /**
-       * @param {AdaptationSet} - AdaptationSet object or string of AdaptationSet
-       *                          id that you wish to set whitelist for
-       * @param {function} - function to filter the representations in
-       *                     order to create your whitelist
-       */
-      setWhiteListRepresentations: function (set, representationFilter) {
-        whiteList = whiteList || {};
-        var setId = set.id || set;
-        if (globalManifest) {
-          var manifest = Object.assign({}, globalManifest.getValue());
-          var currentAdaptation = globalDashManifestModel.getAdaptationForId(
-                                      parseFloat(setId),
-                                      manifest,
-                                      currentPeriodIndex);
-          var type = currentAdaptation.mimeType;
-          type = type ? type.match(/.+?(?=\/)/)[0] : 'video';
-          var currentQualityIndex = globalPlayer.getQualityFor(type);
-
-          var currentRepresentation = currentAdaptation.Representation_asArray[
-            currentQualityIndex];
-          var newIndexInWhiteList = -1;
-          var representations = currentAdaptation.Representation_asArray.filter(
-            representationFilter);
-          for (var i=0; i < representations.length; i++) {
-            if (representations[i].id === currentRepresentation.id) {
-              newIndexInWhiteList = currentQualityIndex;
-              break;
-            }
-          }
-          if (newIndexInWhiteList < 0) {
-            newIndexInWhiteList = selectorFunction(currentQualityIndex,
-                                                   currentAdaptation,
-                                                   representations);
-          }
-
-          whiteList[setId] = representationFilter;
-          globalOgPlayerSetQuality(type, newIndexInWhiteList);
-        }
-        else {
-          whiteList[setId] = representationFilter;
-        }
-      },
-      /**
-       * @param {string} - string value of type
-       * @param {number} - integer value of quality you wish to set
-       */
-      setQualityFor: function(type, value) {
-        if (this.initialized) {
-          if (globalManifest) {
-            var manifest = Object.assign({}, globalManifest.getValue());
-            var mediaInfo = globalPlayer.getCurrentTrackFor(type);
-            var adaptationSet = globalDashManifestModel.getAdaptationForId(
-                                    parseFloat(mediaInfo.id),
-                                    manifest,
-                                    currentPeriodIndex);
-            if(whiteList && whiteList[adaptationSet.id]) {
-              var proposedRepresentation = adaptationSet.Representation_asArray[value - 1];
-              var reps = adaptationSet.Representation_asArray.filter(whiteList[adaptationSet.id]);
-              var i,
-                  len;
-              for (i=0, len=reps.length; i < len; i++) {
-                if (reps[i].id === proposedRepresentation.id) {
-                  globalOgPlayerSetQuality(type, value);
-                  break;
-                }
-              }
-            }
-            else {
-              globalOgPlayerSetQuality(type, value);
-            }
-          }
-        }
-      },
-      getRepresentationsByType: function(type) {
-        if (this.initialized) {
-          if (globalManifest) {
-            var manifest = Object.assign({}, globalManifest.getValue());
-            var adaptation = globalDashManifestModel.getAdaptationForType(manifest,
-                                                                          currentPeriodIndex,
-                                                                          type);
-            return adaptation.Representation_asArray;
-          }
-        }
-      },
-      initialized: false
-    };
-  };
+var WhitelistPlugin = require('./whitelist-ext');
 
   var
     isArray = function(a) {
@@ -293,7 +32,8 @@ var dashjs = require('dashjs');
         }
       }
       return obj1;
-    };
+    },
+    whitelistPlugin = new WhitelistPlugin();
 
   /**
    * videojs-contrib-dash
@@ -338,8 +78,9 @@ var dashjs = require('dashjs');
     }
 
     // Set up plugins.
-    this.whitelistPlugin = new WhitelistPlugin();
-    this.whitelistPlugin.initialize(this.mediaPlayer_);
+    if (whitelistPlugin !== null && whitelistPlugin !== undefined) {
+      whitelistPlugin.initialize(this.mediaPlayer_);
+    }
 
     // Must run controller before these two lines or else there is no
     // element to bind to.
@@ -471,23 +212,23 @@ var dashjs = require('dashjs');
   // Whitelist API
 
   Html5DashJS.prototype.setSelector = function (sFunc) {
-    this.whitelistPlugin.setSelector(sFunc);
+    whitelistPlugin.setSelector(sFunc);
   };
 
   Html5DashJS.prototype.getAdaptations = function () {
-    return this.whitelistPlugin.getAdaptations();
+    return whitelistPlugin.getAdaptations();
   };
 
   Html5DashJS.prototype.setWhiteListRepresentations = function (set, representationFilter) {
-    this.whitelistPlugin.setWhiteListRepresentations(set, representationFilter);
+    whitelistPlugin.setWhiteListRepresentations(set, representationFilter);
   };
 
   Html5DashJS.prototype.setQualityFor = function (type, value) {
-    this.whitelistPlugin.setQualityFor(type, value);
+    whitelistPlugin.setQualityFor(type, value);
   };
 
-  Html5DashJS.prototype.getRepresentationsByType = function () {
-    return this.whitelistPlugin.getRepresentationsByType();
+  Html5DashJS.prototype.getRepresentationsByType = function (type) {
+    return whitelistPlugin.getRepresentationsByType(type);
   };
 
   // Dash.js API
